requirements.txt
PyGitHub
psycopg2